+++ MSAL/src/MSALDeviceInformation.m
@@ -46,6 +46,15 @@ NSString *const MSAL_DEVICE_INFORMATION_SSO_EXTENSION_FULL_MODE_KEY = @"isSSOExt
     {
         _deviceMode = MSALDeviceModeDefault;
         _extraDeviceInformation = [NSMutableDictionary new];
+        
+        if (@available(iOS 13.0, macOS 10.15, *))
+        {
+            _hasAADSSOExtension = [[ASAuthorizationSingleSignOnProvider msidSharedProvider] canPerformAuthorization];
+        }
+        else
+        {
+            _hasAADSSOExtension = NO;
+        }
     }
 
     return self;